--- a/dd-trace-core/src/jmh/java/datadog/trace/core/TracerMapperMap.java
+++ b/dd-trace-core/src/jmh/java/datadog/trace/core/TracerMapperMap.java
@@ -21,20 +21,21 @@ public class TracerMapperMap {
 
   private static final int SPAN_COUNT = 1000;
 
+  private static final TraceMapperV0_4 mapperV4 = new TraceMapperV0_4();
+  private static final TraceMapperV0_5 mapperV5 = new TraceMapperV0_5();
+  private static final CoreTracer tracer =
+      CoreTracer.builder().writer(new LoggingWriter()).strictTraceWrites(true).build();
+
   private final List<DDSpan> spans = new ArrayList<>(SPAN_COUNT);
   private final List<DDSpan> enrichedSpans = new ArrayList<>(SPAN_COUNT);
   private final List<DDSpan> spansWithOrigin = new ArrayList<>(SPAN_COUNT);
   private final List<DDSpan> enrichedSpansWithOrigin = new ArrayList<>(SPAN_COUNT);
 
-  private static final TraceMapperV0_4 mapperV4 = new TraceMapperV0_4();
-  private static final TraceMapperV0_5 mapperV5 = new TraceMapperV0_5();
-  private CoreTracer tracer;
   private MsgPackWriter writer;
 
   @Setup(Level.Trial)
   public void init(Blackhole blackhole) throws Exception {
     writer = new MsgPackWriter(new BlackholeBuffer(blackhole));
-    tracer = CoreTracer.builder().writer(new LoggingWriter()).strictTraceWrites(true).build();
 
     for (int i = 1; i <= SPAN_COUNT; i++) {
       spans.add(createSpanWithOrigin(i, null));