--- a/dd-trace-core/src/main/java/datadog/trace/common/writer/ddagent/TraceMapper.java
+++ b/dd-trace-core/src/main/java/datadog/trace/common/writer/ddagent/TraceMapper.java
@@ -13,7 +13,7 @@ public interface TraceMapper extends Mapper<List<? extends CoreSpan<?>>> {
   UTF8BytesString THREAD_NAME = UTF8BytesString.create(DDTags.THREAD_NAME);
   UTF8BytesString THREAD_ID = UTF8BytesString.create(DDTags.THREAD_ID);
   UTF8BytesString SAMPLING_PRIORITY_KEY = UTF8BytesString.create("_sampling_priority_v1");
-  UTF8BytesString ORIGIN_KEY = UTF8BytesString.create("_dd.origin");
+  UTF8BytesString ORIGIN_KEY = UTF8BytesString.create(DDTags.ORIGIN_KEY);
 
   Payload newPayload();