--- a/dd-trace-core/src/main/java/datadog/trace/core/DDSpan.java
+++ b/dd-trace-core/src/main/java/datadog/trace/core/DDSpan.java
@@ -314,6 +314,11 @@ public class DDSpan implements AgentSpan, CoreSpan<DDSpan> {
   }
 
   @Override
+  public CharSequence getOrigin() {
+    return context.getOrigin();
+  }
+
+  @Override
   public final DDSpan setOperationName(final CharSequence operationName) {
     context.setOperationName(operationName);
     return this;