--- a/dd-trace-core/src/traceAgentTest/groovy/TraceGenerator.groovy
+++ b/dd-trace-core/src/traceAgentTest/groovy/TraceGenerator.groovy
@@ -188,6 +188,11 @@ class TraceGenerator {
     }
 
     @Override
+    CharSequence getOrigin(){
+      return null
+    }
+
+    @Override
     PojoSpan setMeasured(boolean measured) {
       return this
     }