Currently, the tag `_dd.origin` is sent only in the root span. 
This PR adds the internal tag `_dd.origin` to all spans on serialization time. 
We continue storing the `_dd.origin` tag only in the rootSpan.context.

Additionally, we set the fixed origin `ciapp-test` for the spans sent by CI App.

```
java -jar dd-trace-core/build/libs/dd-trace-core-0.81.0-SNAPSHOT-jmh.jar TracerMapperMap -f 1 -wi 3 -i 10 -bm avgt -tu ns
OLD CODE
Benchmark                                      Mode  Cnt       Score      Error  Units
TracerMapperMap.mapTracesV4                    avgt   10  346195.226 ± 2983.504  ns/op
TracerMapperMap.mapTracesWithOriginV4          avgt   10  432187.396 ± 2612.675  ns/op
TracerMapperMap.mapEnrichedTracesV4            avgt   10  546434.042 ± 6559.044  ns/op
TracerMapperMap.mapEnrichedTracesWithOriginV4  avgt   10  627664.892 ± 6165.043  ns/op
TracerMapperMap.mapTracesV5                    avgt   10  128216.576 ± 1400.099  ns/op
TracerMapperMap.mapTracesWithOriginV5          avgt   10  171683.088 ± 2267.104  ns/op
TracerMapperMap.mapEnrichedTracesV5            avgt   10  200731.994 ± 8480.146  ns/op
TracerMapperMap.mapEnrichedTracesWithOriginV5  avgt   10  216636.642 ± 1766.045  ns/op
-----------
NEW CODE
Benchmark                                      Mode  Cnt       Score       Error  Units
TracerMapperMap.mapTracesV4                    avgt   10  322100.408 ±  2237.304  ns/op
TracerMapperMap.mapTracesWithOriginV4          avgt   10  379526.846 ±  3358.523  ns/op
TracerMapperMap.mapEnrichedTracesV4            avgt   10  536631.962 ±  5951.542  ns/op
TracerMapperMap.mapEnrichedTracesWithOriginV4  avgt   10  592819.177 ± 25464.404  ns/op
TracerMapperMap.mapTracesV5                    avgt   10  133245.560 ±  2038.649  ns/op
TracerMapperMap.mapTracesWithOriginV5          avgt   10  149789.677 ±  1484.700  ns/op
TracerMapperMap.mapEnrichedTracesV5            avgt   10  208544.871 ± 14693.354  ns/op
TracerMapperMap.mapEnrichedTracesWithOriginV5  avgt   10  218429.769 ±  3884.491  ns/op
```