@dougqh 2021-06-04
I share the concerns about basing billing on something user provided.

I'll set that aside and focus solely on origin propagation.  To me, propagating origin from one trace segment to another trace segment is logical.

I think ideally; we'd have a TraceSegment object and place the origin on just that object rather DDSpanContext.  However, I thought that was a bit much to ask for this effort.  So for now - for better or worse, this follows the pattern already established for sampling priority.

To deal with partial flushes, we're putting the data on all spans when sending to the trace agent.  I think we all agree that's not ideal.